feat: implementação das regras de negócio

+++ src/main/java/io/github/cursojava/libraryapi/service/AutorService.java
@@ -7,19 +7,29 @@ import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 
+import io.github.cursojava.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.cursojava.libraryapi.model.Autor;
 import io.github.cursojava.libraryapi.repository.AutorRepository;
+import io.github.cursojava.libraryapi.repository.LivroRepository;
+import io.github.cursojava.libraryapi.validator.AutorValidator;
 
 @Service
 public class AutorService {
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository){
+   public AutorService(AutorRepository repository,
+                        AutorValidator validator,
+                        LivroRepository livroRepository){
         this.repository = repository;
+        this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor){
+        validator.validar(autor);
         return repository.save(autor);
     }
 
@@ -28,6 +38,10 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if(possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException(
+                    "Não é permitido excluir um Autor que possui livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
@@ -51,6 +65,11 @@ public class AutorService {
         if(autor.getId() == null){
             throw new IllegalArgumentException("Para atualizar, é necessário que o autor já esteja salvo na base.");
         }
+         validator.validar(autor);
         repository.save(autor);
+    }
+
+     public boolean possuiLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 }
